Edit files in external editor.
With this commit, the user can choose to use xdg-open to edit files.
Inside the snap this is just an wrapper for "snapctl user-open" that
calls the open file portal.

I had to add a "please inform me when you are done" type of dialog
because the call is async. The experience is okay I think.